Move to new file. Build and tests pass.
